rdf-streaming-updater: fix networkpolicy selector
The networkpolicy selector for rdf-streaming-updater should be
selecting the app label of the pod to apply the policy to, not the
chart name.

Since there are two pods in the rdf-streaming-updater release with
different ports being exposed, then two networkpolicies are needed.

Updates service and jobmanager templates to use values for webui port.

Change-Id: I87d2ba7dd5b1967f1f7b84d4489f813cb3594e11