Partially revert "Update cxserver to 2021-03-15-131520-production"
This partially reverts commit 21bf84574df8e381be01f3e436c568cffeb76c5a.

Reason for revert: k8s eqiad was reinitialized today and this version 
was deployed instead of the one that previously was on the cluster. 
Reverting to get the version with the functioning metrics.

Pin the chart version to force using the old one

Change-Id: Ia7e2b875fe36c9deef6731c82361e41eae098a20