admin: Introduce the cluster_group concept
Why:
We started the path for adding this concept in 87fdec37b72b78, which
however was incomplete. Adding a value called cluster_group on the
various environments allows to reference it for including the files on
the next layers of helmfile like calico/coredns/eventrouter. The
namespace/rbac/psp layers are constructed different and don't need this
construct as they aren't layered in the same way

What:
Add a cluster_group value in main.yaml and ml-serve.yaml, which is set
to the same value as the filename (minus the extension). Then use that
value in the 3 helmfiles to include that same file.

Other way it was tried and failed:
* Define the cluster_group value in every environment. However that
  resulted in a construct where the value was multiple times defined in
  the file AND the corresponding file also needed to be included in the
  environments so that namespaces helmfile could use it
* Define the cluster_group value in every environment but not include
  the file in every environment. That ended up in some duplication but
  it did not work as namespaces could not reference their variables

Change-Id: I8760ffcff970f09047f900a54092d3326f32ebb4