linkrecommendation: Use the main_app resources for loaddatasets
Why:
The cronjob resource did not specify a resources section and hence
relied on whatever was configured at the cluster level. In production
clusters that is rather low, leading to the dataset job running rather
slowly

What:
Specify a resources section for the container loading datasets.
Reuse the main_app ones for now until we have some better data

Change-Id: Iaccd3aa818078e00dccb2ec88a9ad792ce990bdb