Envoy: set per_try_timeout for eventgate-main.
According to envoy docs:

> If left unspecified, Envoy will use the global
route timeout for the request. Consequently, when
using a 5xx based retry policy, a request that
times out will not be retried as the total timeout
budget would have been exhausted.

We are seeing 503/504 timeouts in prod for jobs,
so per above, we end up not retrying at all. Let's
try to start retrying.

Bug: T249745
Change-Id: I8989ffa858569e5fd901e65b2b2cfc58c7e62d87