Switch to a different contact email
Switch to contact email noc@wikimedia.org which goes to all SRE
as a contact point for traffic/abuse issues.

The old email was a members-only posting e-mail belonging
to a defunct team and will likely be deleted.

(Also switch to newer underscored user-agent as this is
what is in the default config.yaml in citoid -
the code was kept backwards compatible so this 
old config still works though.)

Bug: T278516
Change-Id: Ie130b56f5a9085d91823051e9d5b412eb491f319