linkrecommendation: Bump limits/requests
Why:
Per T279411, since the switching on of cswiki and bnwiki, arwiki,
viwiki, memory and cpu usage have increased.

What:
Per the task, memory should only be bumped slightly, so bump it to 650MB,
but CPU seems to be throttled quite a bit, so bump it quite a bit more,
to 4.
Bump the requests too, to reflect the average case a bit better, in
order to better inform the scheduler when it does pod placement

For now, leave the chart as is, let's gather more data before updating
it.

Bug: T279411
Change-Id: I283355c36bfb9dac24894223cf308d9c57ec6b2d