Add canary support in scaffolding
Why?
Canarying is the testing in production method where
we release a software change to a small percentage
of our servers, and monitor for potential issues. We
want to be able to support canarying in our applications
running on Kubernetes.

How?
When a release has the "canary" suffix, then kubernetes
will not create a service object for this release but only
a deployment one. Adding "routed_via" to the deployment of
a canary helm release will route traffic to the canary
deployment. The canary helm release must have the "routed_via"
label correctly set.
Also: created a new version of common_templates as this change is
breaking.

Bug: T282148
Change-Id: I64466f9cc8b27a019d92bd9628a314f280893f13