mediawiki: add egress policies to databases
MediaWiki needs to query multiple database sections, and database
types. In some cases, even across datacenters.

To this end, we allow (for now) a pretty lax networkpolicy that
allows to connect to any server in a CIDR on any port between 3306 and 3320.

In the future, we might want to autogenerate a list of CIDRs from
the instances in conftool-data, but for now this is good enough.

Change-Id: I12e90709a79d9d178486966e8a1c85fcf1925449