Add support for knative serving
Knative Serving can be deployed in two ways (from upstream docs):
- kubectl apply -f serving-crds.yaml serving-core.yaml
- Via operators
  (https://knative.dev/docs/install/knative-with-operators/)
The second one seems to be the long term solution, but at the time of
writing upstream considers it in alpha stage and doesn't reccomend it
for production. Moreover, it seems that some dependencies of the
Operator require kubernetes 1.18:
https://github.com/knative/operator/releases/tag/v0.18.1

This is a simple solution to get things moving with Knative and helm, we
may probably want to further refine it in the future (like adding better
segmentation of yaml files etc..).

Bug: T278194
Change-Id: I4be69fd17eaf790cf0e94514ec936442d77fd3e9