miscweb: add a define for the httpd prometheus exporter and use it
So far there are only 2 app types, default or php-fpm.

The php-fpm type includes exporters for httpd and php-fpm.

For this service we want only the httpd exporter, so creating
a new define for just that by itself and then including it.

Also remove the part that only applies to the php app type.

Bug: T281538
Change-Id: Ia18f178701aad589264c38c33338cc20b7c173e5