mwdebug: add servergroup
We want to be able to distinguish mediawiki running on kubernetes from
mediawiki running on physical hardware at the code level.

We've traditionally used the env variable SERVERGROUP to indicate which
group of servers an installation belongs to, and we keep doing this on
kubernetes, but just for pods.

We use a "kube-" prefix to the server group to identify running on
kubernetes.

Bug: T284418
Change-Id: I39d331f0479c205167e23b3240460c8b6f6643db