mwdebug: Allow non-roots to perform a rolling restart
With our RBAC setup, only root users and tiller itself are able to delete
pods. So for a normal deployer, the only way to cause a rolling restart
at the moment is to add a recreatePods: true in their helmfile.

That is very inconvenient, imho. This patch gives us the ability to
perform a rolling restart by controlling behaviour with a command-line
flag, --state-values-set roll_restart=1.

Change-Id: I7d25a6690466ffb101900481169d70f2fd8266c3