helmfile: allow performing a rolling restart
With this change, we expand the ability to perform a rolling restart of
all pods in a deployment by using a runtime switch in helmfile.

With this change, helmfile will perform a rolling restart if ran with

helmfile -e eqiad --state-values-set roll_restart=1 sync

Please note: helm deprecated the usage of --recreate-pods with version 3,
which we're transitioning to soon, but it will still be available for
quite some time.

While I understand the reasoning behind the deprecation, using alternative
systems to obtain similar results (like a random annotation to add to the
deployments) would work less well for small deployments like most of ours -
namely it will always imply a rolling restart not only when needed, but
also the next time someone runs helmfile (as the random annotation will
disappear).

Change-Id: Id96c98f2a4bbbf9c28844725a59c7c201646f2ff