Kubeflow: fix secret in chart and update helmfile
The Secret config in the kubeflow-kfserving chart needs to be b64
encoded to be valid.
Moreover, the kfserving secrets are stored under a more precise
/private path on deploy1002, so changing helmfile accordingly.
Last but not least, remove the namespace setting from helmfile,
since it clashes with the namespace definition (kfserving-system)
listed in the chart (helm fails to deploy due to multiple attempts
to create the same namespace).

Bug: T272919
Change-Id: I462434c5bca7b5c953d74ac760d748e7552fdbaf