shellbox: Disable php-fpm slowlog
slowlog is not useful for Shellbox because nearly all of the execution
time is being taken up outside of the PHP code. Plus it requires the
ptrace capability, which we'd like for Shellbox to not have.

Per <https://www.php.net/manual/en/install.fpm.configuration.php> setting
the timeout to 0 disables the functionality.

Bug: T288315
Change-Id: I4a813ed2236adab7c8cb5591574b3b32be39e951