linkrecommendation: use CNAME for analytics-web
The corresponding CNAME lives in
https://gerrit.wikimedia.org/g/operations/dns/+/12acd26316e6de82d5979d1a77d5e9f21e3a970f/templates/wmnet#648.
This should effectively be a no-op.

We're switching to the CNAME because as part of T285355 we'll be cutting
over from thorium to an-web1001 in the near future, so we want to be
able to just change the cname.

Additionally, this sets up the new egress for an-web1001. When we
cutover fully from thorium to an-web1001, we'll want to circle back to
remove the thorium-specific egress entries.

See https://netbox.wikimedia.org/dcim/devices/3428/ for where the ipv4
and ipv6 values for an-web1001 came from.

Bug: T285355
Change-Id: I08c0108640e310b943b9ebaebd76959a24609eba