api-gateway: allow /staging/ testing namespace only in staging
We need to provide a more free-form way for services to be initially
onboarded to the gateway. To this end, add a /staging/ endpoint that
allows arbitrary services to be defined in advance of moving them to
somewhere more appropriate in the gateway namespace hierarchy.

This namespace is only created when configured on staging. Configuring
staging namespace entries in production should have no effect.

Also fix a minor config formatting issue where we were defining an
endpoint-level config for discovery services which was creating an
empty cluster that would never be used.

Bug: T289583
Change-Id: If9fe4186e75fdb9dc8bc8ccc3f0a392329efe4dc