helmfile.d admin add dedicated deploy user
This change adds ClusterRoles and RoleBindings for a dedicated deploy
user for helm3 deployments. The user has the same permissions like tiller.
The introcution of a new user was discussed here:
https://phabricator.wikimedia.org/T251305#7314778.
I also introduced two optional values. readOnlyUser overwrites the
name of the user, similar to the deployUser value. deployClusterRole
overwrites the ClusterRole for the deploy user, similar to
tillerClusterRole

Bug: T251305
Change-Id: Iad5c329899a987a5d519868648c506e69daa4df8