knative-serving: add secrets chart to helmfile config
This change is needed to allow the knative-serving helmfile
config to be able to push a TLS certificate secret to
the istio-system namespace, so that it will be picked up
by the istio ingress gateway. Knative is responsible for L7
configs for istio, that also includes HTTPS.

The change also add a basic TLS config for port 443, to
test the new functionality on ml-serve-eqiad.

To make everything works we also need to create the
'inference-tls-secret' k8s Secret config in puppet private.

Bug: T289835
Change-Id: Ic78b4efcfc06b32ca578933c136e55e6588da8c8