Add revscoring-editquality as first ml-service to helmfile.d
This is a proposal for a new helmfile.d directory called
"ml-services", aimed to host configs for all the services that
will run on ml-serve clusters.

These services are split into a separate directory to cleanly split
what services belong to the "main"/staging clusters and what to
the ml ones.

The Kubeflow Kfserving layer (already deployed via admin_ng) allows
the definition of the following within the same namespace:
- an InferenceService resource (the CRD definition is provided
  Kubeflow) that represent the ML model to deploy with its config
  details.
- a Secret to hold credentials to fetch models from Swift (we are
  using Thanos Swift for the moment, but we'll migrate to MOSS once
  it will be ready).
The above need to be deployed using a service account able to spin
up pods and read Secrets stored in the namespace. For the testing
phase we have used custom/manual configs:
- a custom service account
- a RoleBinding for the above account with allow-privileged-psp.

The upstream docs are available here (to have a better overall
idea):
https://github.com/kubeflow/kfserving/blob/master/docs/samples/storage/s3/README.md

The assumption of the change is that the `deploy` user will be
picked up and that it will be able to satisfy the above constraints.

The helmfile config uses the recent `secrets` chart that is needed
to deploy custom Secret configs picking them up via helmfile private.

Last but not the least, some details about how we plan to deploy
models. Our first target is to host all ORES models on ml-serve,
that are around 100, split into 4 categories:
{article,draft,edit,topic}quality
In this change we want to deploy one model for the editquality category,
(revscoring is the name of the library that use on top of scikit-learn,
but also a nice word that represents Revision Scoring models), that
is the one for enwiki. If this change is sound the plan is to add more
when available on Swift.

Bug: T286791
Change-Id: I4463dc393ce8cfb5285ff967bd7341c742341b37