mediawiki: allow rewriting static assets to multiversion
Given in kubernetes we've separated the webserver from the php executor,
we can't directly access static assets under /static/current
(which used to be a symlink). This is a first attempt to rewrite those
urls to be directed to /w/static.php, and thus be served by the php engine.

Bug: T285232
Change-Id: If644b7930247a7390fe8d98aa475b6824eecd896