toolhub: set "docker.pull_policy: Always"
Tell our charts to tell Kubernetes to always check with the docker
registry to see if the hash for a tag is present before pulling instead
of just checking to see if the tag itself is present
("docker.pull_policy: pullIfNotPresent").

The issue this is intended to address is the use of "latest" tags for
things like prometheus-mcrouter and prometheus-mcrouter-exporter. Once
a Kubernetes node has pulled *any* :latest tag for an image
'pullIfNotPresent' will happily use that image. For a tag like "latest"
or "0.x" that may be updated over time this can lead to difficult to
debug behavior where "latest" on the exec node is not the same as
"latest" in the registry or "latest" in a development environment.

The main alternative to this would be to pin each and every container
used in the deployment to an immutable tag like "0.41.0-4-20210718" or
"0.0.1-2-20210919". The main downside of this is that these tags would
need to be manually updated in the values.yaml to keep tracking useful
changes introduced in the sidecar images. That may or may not turn out
to be a better idea, but I'm going to try this as a "quick fix".

Bug: T291442
Change-Id: Ie37c6204fba552f3396ddde3748f880d121ab37f