Eventgate TLS proxy: lower local idle_timeout
For remote service we usually set custom idle_timeout via
keepalive, but for local service (envoy for inbound
connections) we almost never do. The default is 5s, which
is exactly the default for nodejs keepAlive, so we get a race
when sometimes by the time the request reaches node, the
keepAlive has already expired.

Lower the timeout for eventgate to 4.5s.

Bug: T215001
Change-Id: I633724da450726d932e0036de9dd2e9870bbedbf