Lower the default TLS proxy inbound idle timeout
Most of our services are implemented in node.js,
where the default keepAlive timeout is 5s. Also,
most services don't override idle_timeout for
local connection. So we get a race condition where
sometimes by the time inbound request reaches the
service, it's keepAlive has already expired.

I propose to lower the default timeout a little so that
everything works out of the box for the most common case.

Change-Id: Iae3e52b41c02a14c9007c96153a23956e6f8522e