rdf-streaming-updater: Reduce flink mem
The container is still being killed from time to time, esp when doing
savepoints. I suspect the limits that flink imposes to itself to not
be very accurate causing the OOM killer to be triggered.
Increase the margin between container hard limit and flink limits to
mitigate this.
The JVM metrics appear to be very sane with almost no old gen counts
and very few young gen counts.

Change-Id: Ia0ac06a9594cd91cd43ab8c11006fa381462f3aa