kubeflow-kfserving-inference: create sa to read s3 secrets
The kfserving s3 upstream tutorial suggests to use a dedicated
service account to be able to read the s3 credentials in the
related secret (in our case, to access Swift).
This change adds a basic service account to the chart, and
its base config to the revscoring-editquality service.

Add also a link to _helpers.tpl to be able to add helm-related
labels to the ServiceAccount created by the chart.

Upstream docs:
https://github.com/kserve/kserve/blob/master/docs/samples/storage/s3/README.md

Change-Id: I77ecf99b404db3a07b98eb58883d3104d42015af