toolhub: set https_proxy envvar
A big "duh" moment for the author. Manual testing has shown that we need
both http_proxy and https_proxy set in the environment. Python's
`requests` library uses the first for urls using the http protocol and
the second for https protocol urls. Previously it had been naively
believed that http_proxy would be used for both non-TLS and TLS
connections if https_proxy was not set.

Bug: T292027
Change-Id: I275bdd3e4b9ec84525860e4e1b29f9127b7d558e