toolhub: Set concurrencyPolicy=Replace for CronJob
A CronJob with sidecar containers is a bit of an ugly mess for
Kubernetes to deal with. The main job container's exit will not force
the sidecar pods to exit, so the whole job stays "alive" forever as far as
Kubernetes is concerned.

We will use a sneaky trick of setting the job's concurrencyPolicy to "Replace"
to overwrite the stuck job with a new job each time the scheduler fires.
This will leave envoy and mcrouter containers running all the time that
are not technically needed, but does not require us to architect a way
for the main job container to signal the sidecars to terminate using
something like `kubeexit`.

See also
<https://medium.com/finnovate-io/how-to-prevent-kubernetes-cron-jobs-with-sidecar-containers-from-getting-stuck-912c0f1497a3>

Bug: T292027
Change-Id: Iace8fdf411cd840e3efa2eeef8dc58a20018ed9e