toolhub: Bump container version to 2021-10-12-152757-production
Update to container containing the contrib/cronjob/crawler.sh helper
script needed by the CronJob in the v1.0.0 Toolhub chart.

Tracking these dependencies is tricky. It is pretty easy to make
a change to the chart be dependent on a change in the codebase using
gerrit's `Depends-on` syntax. Making these helmfile changes properly
dependent on the chart and the code is less obvious, especially since we
need the chart change and the helfile change to be submitted as separate
gerrit change sets to make the CI diff job work as expected.

Bug: T292861
Change-Id: I9c1fb1031efa8d1c856b4ec23a5f525fe16ec981