Raise the priority of wikibase-InjectRCRecords job
This job ensures that users see recent changes from wikidata in a timely
manner in their watchlist for the Entity changes that affect the pages
that they are watching. This is important, because that way these users
can spot vandalism quickly that affects the pages that they care about.
Currently, this tends to be crowded out whenever someone edits a
template that is included on a lot of user pages, because then
LocalGlobalUserPageCacheUpdateJob is triggered which monopolizes all the
resources.

This job is typically injected less than 10 times per second and has a
mean duration of less than 0.3 seconds, so a concurrency of 4 should
give us enough breathing room for spikes and recovering from unusual
backlogs.

Bug: T293385
Change-Id: Ie3e6b664af3e8f63fba07cc5ceff4155f5d4db1c