Increase concurrency for wikibase-InjectRCRecords job
In 9d5d1afd49a90186b41f0a47f756c181d78fc113 we raised the priority of
the job, which improved things, but the backlog is still increasing and
we're seeing a high 99th percentile job run time of 3+ seconds.

So, increasing the concurrency to be able to better deal with that.

Bug: T293385
Change-Id: Ibc8997f9e5576ef9f6e01c2c9c7253395339f90c