Increase concurrency for EntityChangeNotification job
This job has suffered recently an increase in job run duration, the
cause of which is not yet clear.
Also, together with the current mid-day maximum in job insertions, we're
seeing a growing backlog that affects the "between change on wikidata
made and injected on client wiki" Service Level Objective (SOL).

Change-Id: I60e3e16f142e441f0977bf12820af44ec24d5ac5