api-gateway: improve configuration naming for discovery_endpoints
Naming change to improve configuration of discovery_endpoints.
When a new enpoint is added, its name become part of a cluster
named "endpoint_cluster". For the current linkrecommendation
service, it would be:

linkrecommendation.discovery.wmnet_cluster

The name can be re-used in pathing_map for example (to rewrite
URIs or Host headers) but it is not very nice to read at first
sight. This change introduces a new "host" parameter (that by
default joins the endpoint name with .discovery.wmnet) and
decouple it from the endpoint name.

Bug: T288789
Change-Id: I4751523726890801ae3d671f6a153940e4541cbd