kserve-inference: improve labels and network policy rules
The current network policy label matching doesn't work
(since the previous code change for the chart was wrong,
there is a mix of labels and annotations) and it seems
more convenient to re-use the app/version labels that
the other chart are using for consistency and flexibility.
We cannot use the 'app' label since it seems used by
knative-serving (it sets the label equal to the knative
revision name), so this change introduces the 'app-wmf`
label.

The network policies also need a specific rule to open
the port of the sidecar container called 'queue-proxy',
that knative-serving uses when the "real" kserve
container is not ready.

Bug: T289834
Change-Id: I3cad05fe63309dd250d7897916fb9c5b957d5e25