knative-serving: add basic egress network policies
The knative serving pods are various controllers that need
to fetch data from the K8s API. This change adds the basic
egress policies (ingress still WIP) to allow the pods
to work correctly.

One of the controllers (called 'controller') needs also
to fetch data from the docker registries, to convert
docker image tags to digests.

Some wmf-specific labels are also added to improve
consistency with the rest of the charts.

Bug: T289834
Change-Id: I50dc9a46f94f1391a14fc15457687890e0351560