helmfile.d: move Docker registry's IPs to ml-serve.yaml
The Knative controllers need to reach the Docker registry api
to convert image tags to digests. The previous change naively
added the docker registry's codfw service IP to ml-serve-codfw.yaml,
and the eqiad one to ml-serve-eqiad.yaml. With a discovery endpoint
it may happen (like now) that the eqiad cluster needs to fetch from
the codfw service endpoint (since the eqiad one is depooled).

Bug: T289834
Change-Id: I4bce9334cac03914162110db93dd4757cda1d47f