kserve-inference: improve network policies
Changes:
1) Temporarily remove the link to default-network-policy-conf.
since it is too broad for the ML use case. A follow up with
ServiceOps is needed to figure out how to better share the file.
The egress rules for the MW API have been copied to the
revscoring's helmfile config.
2) Allow traffic to port 9090, since some knative pods need to
fetch prometheus metrics to do their work (like the autoscaler).

Bug: T289834
Change-Id: Ib7b42be8e1d33bd4f582f374d07fc96b14fe38e0