Skip (re-)building helm dependencies during CI
With the changes of using local charts for CI[1], helmfile would
build dependencies over and over again. Leading to quite an increase
in CI runtime. With --skip-debs, helmfile skips this step.

This requires a charts dependencies to be checked in to git (the
tarballs under <chart>/charts/*.tgz) which is verified by every
task running "helm template" already (:helm_diffs, :validate_template).

[1] 605cf0c1fdcc710c0da43101644b5f829166013c

Change-Id: I074b9aa4367147571c9d10ed343f1d3014b0178c