charts: fix affinity indentation in charts and scaffold chart
This change fixes the identation for affinity in deployments. With
the current indentation it is not possible to set the affinity without
generating invalid yaml:

[ERROR] templates/deployment.yaml: unable to parse YAML
error converting YAML to JSON: yaml: line 24: mapping values are not
allowed in this context

I also added a example to _scaffold on how to use affinites (because
nodeSelectors and nodeAffinity was used differntly in some charts).
Furthermore I used nindent instead of indent in all charts to keep
the yaml template formatting.

The issue was fixed in other charts individually (see
I80067918f3a7ea817d8dbaebb7d5cf16734cbe72 or
I687f0ff49ef9aaf6889abf5bf1cdbe62997909bc).

Change-Id: Ifb582b0105137a47aa94abe29e4df3a9ae58a8c3