Update termbox to 2021-11-26-093451-production
Includes a part of the (not yet finished) migration to make Termbox
ready for Vue 3. Should not introduce any functional changes, but if
it’s broken for some reason, we’d like to know about it before finishing
the migration.

Also pulls in other changes since March, such as dependabot updates. I
can’t say if any of these affect SSR. (The tagging changes shouldn’t
have an effect on SSR.)

Bug: T296202
Change-Id: I32c6d6be7e5df1fcb1b79a7655dfb6da724e2669