admin_ng: refactor istio helmfile config to allow egress gateways
The ML team is experimenting with Istio Egress gateways, that have
basically the same configs as the Ingress ones, but they also need
to be allowed to contact specific APIs (Swift, MW API, etc..).

The idea is to generalize the current helmfile_istio-ingress.yaml
file (and related parameters to enable/disable it) to allow
the definition of egress gateways.

A basic config is also added for the ml-serve clusters, namely
a single egress gateway deployed in the istio-system namespace.
The rationale for this choice is that all services in ml-serve
share the same structure and CRDs (provided by Kserve), so
having alternative layouts like, for example, one egress gateway
for each namespace is a little overkill.

The Istio Gateway specific config bits (for layer 7 routing etc..)
will be added to the knative-serving's chart (since it currently
holds also the ingress ones).

Bug: T294414
Change-Id: I2315622dd33892e83e3881618d588f800801f434