knative-serving: add support for istio egress gateway
The Istio ingress configurations are added/configured by Knative,
so it seems good (as initial step) to add the egress gateway bits
in the same chart (since we don't have any chart dedicated to istio).

The ML team would like to test the Istio Egress gateway instead of
using the envoy sidecar pods, to be more consistent with
the Istio+Knative stack and be more future proof (for example, if
we turn on istio mTLS mesh in the future).

Change-Id: I9d5d08c9091418d9d58bb9e89f5a55617f6ac9b6