Rakefile: remove helm2 from Rakefile, bump scaffold to v2 api
This change removes the usage of helm2 binary inside of the Rakefile
tasks. Since we are running helm3 only we shouldn't lint and use helm2
in CI anymore.
The logic around Chart API v1 and v2 will remain in the Rakefile
until T295750 is closed. I addded a comment to make that clear.
Rake tasks will fail now when helm2 is used.

Bug: T251305
Change-Id: If9c8026230a323a81d09e543b594d8e307ad0999