mediawiki: add the ability to inject php files for debugging
We often need to debug things in production where running a hand-crafted
script in production allows us to find out a lot of information faster.

Add a mechanism to inject small php files under /w/debug using a
configmap to the chart.

Also enable the debug system for mwdebug, where it will be mostly used.

Bug: T297613
Change-Id: Ia6098463f694bf8bb30f76cb2ccd508a857434dd