mediawiki: add ability to inject apache configurations early
We already can inject configuration late using mw.sites. For debugging a
current task, we need to test different configurations of apache and
rebuilding the image chain for every test is decidedly an overkill.

So, allow users of the chart to inject values early in the configuration
that would on the long run probably be better served from the base images.

Also for the bug at hand enable the remoteip module in mwdebug

Bug: T297613
Change-Id: I9a04d214a2fb0eeec120edd505a1d12490e6d3a1