Expand helmfile_namespace_certs to support the ml-serve use case
The ml-serve clusters are currently using a single TLS certificate,
issued by the Puppet CA, for the Istio Ingress Gateway config:

inference.discovery.wment

The certificate is coupled with the knative-serving chart, since
all the Gateway/etc.. Istio configs are contained in it.
In order to re-use as much code as possible, it would be
convenient for the ML team to use helmfile_namespace_certs
for the knative-serving namespace, but it is currently not
possible since there is a specific flag that skips system namespaces.

The idea of the change is to introduce a flag, by default
behaving like the current code, that would allow the ML team to
set the tlsHostname config only for the knative-serving namespace
(and turning it off for the other ones).

Bug: T298976
Change-Id: I0b9ee27c6839940cddb9d339df0236b7ccf04219