Refactor Rakefile
Transform the bowl of repeated logic that is the current Rakefile
into an ordered set of classes.

Code is divided in:
* Runner classes, that find Assets to test, and exectue tests on them
* Asset classes, that actually perform the tests on the individual Assets
* Outcome classes, that collect the results of tests
* View classes, that offer a view for the user.

The new code takes a few different approaches that should make it
slightly faster than the old one.

On my computer(TM):
* tasks lint, helm_diff and validate_template took 61 seconds
* task check_charts took 35 seconds
* tasks validate_deployments and deployment_diffs took 30 seconds
* task check_deployments took 35 seconds
* task check_admin took 39 seconds
* tasks admin_{lint,validate,diff} took ~ 1 minute

deployment checks are slighlty slower because we actually do a small
amount more work since we're not also validating them, thus we don't
save one sweep of helmfile template in that case.

Also added tests at least for BaseAsset's main functions.

Change-Id: Ibdb794dbb4a52a249c95196aac22096a962acc43