charts: remove depricated helm test annotation, fix hook-delete-policy
This change removes the depricated helm2 hook annotation "test-success" and replaces
it with the helm3 alternative "test".

The management of test resources also changed with helm3. So this change adds the
hook-delete-policy "hook-succeeded" to the default policy "before-hook-creation".
The combination of both policies means, successful test pods are cleaned up and
all previous test resources are cleaned up when a test is executed.

Bug: T251305
Bug: T276949
Change-Id: I9a28bcd0ff3549688028e446b1655707d2d6787f