changeprop-jobqueue: increase CPU and memory allocation
Currently jobqueue jobs are failing to reach their concurrency in many
jobs, leading to backlogs in many places. We've temporarily increased
concurrency for these jobs, but increasing CPU is most likely a safer
bet. We saw similar issues with regular changeprop in the past and
increasing CPU resolved many of the issues then, and it is perhaps
oversight that jobqueue itself didn't have its cpu limits
increased. We are seeing some throttling in jobqueue in a similar
fashion, even if the impact is at present much more subtle than the
issues changeprop (and the larger cluster) historically suffered when
hitting CPU and memory limits.

Bug: T300914
Change-Id: I24931e2f39c42ee8e5df2bc41fe1b7c27c0c471b