changeprop-jobqueue: decrease CPU use, increase number of nodes
Increasing CPU and memory reduced CPU throttling, which is a good
thing. CPU use is not up, which makes me think that 2 is a little high
for our purposes. However, given how jobs are distributed to
changeprop instances, our main constraint is actually which pods pick
up which jobs - some pods have more jobs than others and get
oversaturated meaning we still experience issues with
concurrency. Currently we don't have a smarter way to distribute jobs,
(although this is being looked at) but for the short-term increase the
number of instances so the distribution will hopefully be a little bit
less uneven.

Bug: T300914
Change-Id: I9ce8d49d38745c996325cfe511f670b7422418e6