kserve-inference: add configuration for revscoring models
The idea is to add a separate config file for revscoring
models, so that we can configure isvc resources with
more DRY set of settings. The new dedicated config
has a name, `revscoring_inference_service`, to
avoid confusion with the generic `inference_service`
one.

The STORAGE_URI variable can be refactored as well,
but due to how we build an isvc now, it may take more
time. We can improve that part later on if we feel
the need.

Change-Id: I1116f867cd6a3a4399c298febcff21e081a3a595