Rakefile: enhance output on broken yaml/new assets
When the validation step generates an invalid yaml, helm exits with a
rather uninformative error message. We should instead use the --debug
switch for helm to output a yaml even when it's not a valid one, so that
we can validate it in software and extract a more informative error
message.

Also add again the validation report to the deployments report - while
it's not run by default, it can be run on-demand.

Finally, when generating templates on the original directory,
re-compute the fixtures.

Change-Id: I07fa355815afbeda5d31f3a1549b5886ff16b8ae