mediawiki 0.2.0: Add mw.localmemcached.enabled value
The default value is false.  If set to true, one or more memcached
containers will be started in the same pod as the app and httpd
containers.  These two containers can serve as stand-ins for
production components such as nutcracker and mcrouter.  The purpose of
this value is to enable image testing in non-production environments
such as train-dev.

Change-Id: I936a8338f04984a6681d7e372eaf6009a78411e3