jobqueue: set CPU request
Given that jobqueue is a fundamentally bursty service, CPU limits are
most likely causing the service to be extensively throttled even
though it's not hitting its quota in terms of overall CPU usage. This
might be one of the sources of our issues with concurrency - add a
higher request to see if this has some impact.

Bug: T300914
Change-Id: Ib2e56132280f8e1fc1abcffcbd3d7021d7b94aaf