Move ml-serve pod configs to Istio proxy sidecars
The idea is that every pod in the mesh will be able to contact
egress endpoints, like what the ServiceOps team does in the other
prod clusters. The circuit breaking settings are tuned to be applied
to the mesh proxies, so it makes sense to simplify the ml-serve
clusters config to avoid an egress gateway (at least for the moment).

This change:
- Allows only some namespaces to get inspected by istio-cni (to inject a sidecar)
  using a specific label.
- Allows pods to contact Thanos and the MW API (since all sidecars will
  need to).
- Adjusts some istio-specific configs to work with the new
  mesh settings.

Bug: T297612
Change-Id: Ib99f7387ff4b39550def8064faa2c8b0f9bce69a