Add the networkpolicy for the setups as a pre-install hook
Currently the three setupjobs attempt to run before the subcharts are
deployed. However they cannot access the network resources that they
need because the networkpolicy has not yet been deployed.

This commit addresses that situation by creating the network policy at
an earlier stage in the deployment, along with the secrets.

Bug: T301454
Change-Id: Iaa583ee5be22f22258a0c10f7febde73c8e40827