Enable SSL/TLS for accessing the datahub-gms service
The datahub-frontend component makes use of an environment variable
DATAHUB_GMS_USE_SSL in order to determine whether or not to use SSL/TLS
when accessing the GMS service.

Since we use an envoy TLS proxy for the GMS service we need to enable
this for all ouf our deployments.

Bug: T301454
Change-Id: I7315381e2047af2ec10f05b59149ee83da666731