Configure LDAP authentication for DataHub
This commit introduces an optional ConfigMap object that is mounted as a
JaaS configuration file to the datahub-frontend component when LDAP
authentication is enabled.

The current configuration allows anyone with a valid LDAP account to
access DataHub. I tried to filter based on group membership, but so far
have not got it to work.

Bug: T301462
Change-Id: I21ebcfe35a6747562bf07f65e1f96b4b94931037