Add calico BGP peering settings for ml-serve100[5-8]
The new ml-serve-eqiad nodes are in row E/F, that have a different
network configuration from the rest of the racks. They are switches
doing layer 3 routing, so the calico pods need to establish
a BGP adjiacency with the ToR's IPs instead of the cr{1,2}-eqiad
loopback addresses (like we do for the rest of the nodes,
connected to layer2-only switches).

Bug: T306649
Change-Id: Id70223e976aacf467eff5e975c6b3122121ab08a