Double the number of eventgate_analytics_external replicas
We have recently doubled the number of workers from 12 to 24 in order to
try to deal with a latency in the 90th and 99th percentiles. The
previous change had an effect, but it was subtle. This change doubles
the number of workers again to see if the reduction in latency continues
as replica numbers increase.

Bug: T306181
Bug: T294911
Change-Id: I18aa998de6eaf554851f16a83138fc3a53b5de71