service-proxy: Set SNI and Host header for ingress services
Kubernetes Ingresss (istio) requires SNI and Host header to be set in
order to be able to serve the correct certificate and route the request
properly.

Set SNI and Host header in service-proxy (envoy) config for services
in profile::services_proxy::envoy::listeners that have "use_ingress"
set to true.

Bug: T312225
Change-Id: I1c430f4d2b50d93dc5a56c754c555663483f6826