shellbox: Bump replicas to 24 to support request spikes
The change was actually made live on 2022-07-03 during a production
incident, so this commit is actually a noop for production, and just
catches the charts repo up to reality.

After evaluating the impact of I3ce297fc39bcfb3a74975c0d98800ac3a0b32fb4
we might revert this, but for now let's have a clean diff.

Bug: T310557
Change-Id: I30d95940f23a7ec1f84c22dbd8eed0f4775f7790