changeprop-jobqueue: increase number of workers to 2
Currently we're seeing insufficient concurrency in many jobqueue
jobs. Increasing number of workers to see if we get any concurrency
boost in existing nodes. This approach was suggested by Petr before he
left as an avenue to explore so let's see how it behaves. This value
was set to 1 as part of the kubernetes migration.

Bug: 300914
Change-Id: I3e851dc2e1fc58163936dec2ce801615850cedc8